feat: new functions in module
